Training Loop Works
Training loop works now. But loss is not decreasing.

We suspect this is because we are only training for 1 batch of batchsize 4. We need a lot more images to train on. Also the number of epochs we are training for is minimal, so this needs to be increased as well.

Also need to figure out how to run this with GPU instead of just CPU.